Implement modules page.rs and book.rs from models.rs
