+++ page-hunter/src/page.rs
@@ -1,21 +1,17 @@
 use std::fmt::{Debug, Display};
 
-use super::errors::{ErrorKind, PaginationError};
-
 #[cfg(feature = "serde")]
 use serde::{
     de::{Deserialize as DeDeserialize, Deserializer as DeDeserializer, Error as DeError},
     Deserialize, Serialize, Serializer,
 };
-
 #[cfg(feature = "utoipa")]
 use utoipa::{
-    openapi::{schema::Schema, ArrayBuilder, KnownFormat, ObjectBuilder, SchemaFormat, SchemaType},
+    openapi::{schema::Schema, KnownFormat, ObjectBuilder, SchemaFormat, SchemaType},
     ToSchema,
 };
 
-/// Result type used throughout the library for result handling.
-pub type PaginationResult<E> = Result<E, PaginationError>;
+use crate::{ErrorKind, PaginationError, PaginationResult};
 
 /// Model to represent paginated items.
 ///
@@ -97,7 +93,7 @@ impl<E> Page<E> {
             false => self.get_total().div_ceil(self.get_size()).max(1),
         };
         if expected_pages.ne(&self.get_pages()) {
-            return Err(PaginationError::from(ErrorKind::FieldValueError(format!(
+            return Err(PaginationError::from(ErrorKind::InvalidValue(format!(
                 "Total pages error: expected '{}', found '{}'",
                 expected_pages,
                 self.get_pages(),
@@ -106,7 +102,7 @@ impl<E> Page<E> {
 
         // page must be less than pages - 1.
         if self.get_page().gt(&(self.get_pages() - 1)) {
-            return Err(PaginationError::from(ErrorKind::FieldValueError(format!(
+            return Err(PaginationError::from(ErrorKind::InvalidValue(format!(
                 "Page index '{}' exceeds total pages '{}'",
                 self.get_page(),
                 self.get_pages(),
@@ -115,7 +111,7 @@ impl<E> Page<E> {
 
         // if page is less than pages - 1, items length must be equal to size.
         if self.get_page().lt(&(self.get_pages() - 1)) && items_length.ne(&self.get_size()) {
-            return Err(PaginationError::from(ErrorKind::FieldValueError(format!(
+            return Err(PaginationError::from(ErrorKind::InvalidValue(format!(
                 "Items length '{}' is not equal to page size '{}' for an intermediate page '{}'",
                 &items_length,
                 self.get_size(),
@@ -129,7 +125,7 @@ impl<E> Page<E> {
                 .get_total()
                 .ne(&((self.get_pages() - 1) * self.get_size() + items_length))
         {
-            return Err(PaginationError::from(ErrorKind::FieldValueError(format!(
+            return Err(PaginationError::from(ErrorKind::InvalidValue(format!(
                 "Total elements error: expected '{}', found '{}'",
                 (self.get_pages() - 1) * self.get_size() + items_length,
                 self.get_total(),
@@ -143,7 +139,7 @@ impl<E> Page<E> {
         };
 
         if expected_previous_page.ne(&self.get_previous_page()) {
-            return Err(PaginationError::from(ErrorKind::FieldValueError(format!(
+            return Err(PaginationError::from(ErrorKind::InvalidValue(format!(
                 "Previous page index error: expected '{:?}', found '{:?}'",
                 expected_previous_page,
                 self.get_previous_page(),
@@ -157,7 +153,7 @@ impl<E> Page<E> {
         };
 
         if expected_next_page.ne(&self.get_next_page()) {
-            return Err(PaginationError::from(ErrorKind::FieldValueError(format!(
+            return Err(PaginationError::from(ErrorKind::InvalidValue(format!(
                 "Next page index error: expected '{:?}', found '{:?}'",
                 expected_next_page,
                 self.get_next_page(),
@@ -457,174 +453,5 @@ where
 				)
        		).into()
 		)
-    }
-}
-
-/// Model to represent a book of paginated items.
-/// #### Fields:
-/// - **sheets**: Represents the ***sheets*** in a [`Book`] as a [`Vec`]  of [`Page`].
-pub struct Book<E> {
-    sheets: Vec<Page<E>>,
-}
-
-impl<E> Book<E> {
-    /// Get ***sheets***
-    pub fn get_sheets(&self) -> &Vec<Page<E>> {
-        &self.sheets
-    }
-
-    /// Create a new [`Book`] instance.
-    ///
-    /// ### Arguments:
-    /// - **sheets**: A reference to a [`Vec`] of  [`Page`], where `E` must implement [`Clone`].
-    ///
-    /// ### Returns:
-    /// A [`Book`] if successful, otherwise a [`PaginationError`] is returned.
-    ///
-    /// ### Example:
-    /// ```rust,no_run
-    /// use page_hunter::*;
-    ///
-    /// let sheets: Vec<Page<u32>> = vec![
-    ///     Page::new(&vec![1, 2], 0, 2, 5).unwrap_or_else(|error| {
-    ///         panic!("Error creating page model: {:?}", error);
-    ///     }),
-    ///     Page::new(&vec![3, 4], 1, 2, 5).unwrap_or_else(|error| {
-    ///         panic!("Error creating page model: {:?}", error);
-    ///     }),
-    /// ];
-    ///
-    /// let book: Book<u32> = Book::new(&sheets);
-    /// ```
-    pub fn new(sheets: &Vec<Page<E>>) -> Book<E>
-    where
-        E: Clone,
-    {
-        Book {
-            sheets: sheets.to_owned(),
-        }
-    }
-}
-
-/// Implementation of [`Clone`] for [`Book`].
-impl<E> Clone for Book<E>
-where
-    E: Clone,
-{
-    fn clone(&self) -> Self {
-        Book {
-            sheets: self.sheets.to_owned(),
-        }
-    }
-}
-
-/// Implementation of [`Debug`] for [`Book`].
-impl<E> Debug for Book<E>
-where
-    E: Debug,
-{
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "Book {{ sheets: {:?} }}", self.sheets)
-    }
-}
-
-/// Implementation of [`Default`] for [`Book`].
-impl<E> Default for Book<E> {
-    fn default() -> Self {
-        Self { sheets: Vec::new() }
-    }
-}
-
-/// Implementation of [`Display`] for [`Book`].
-impl<E> Display for Book<E>
-where
-    E: Debug,
-{
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "Book {{ sheets: {:?} }}", self.sheets)
-    }
-}
-
-/// Implementation of [`IntoIterator`] for [`Book`].
-impl<E> IntoIterator for Book<E> {
-    type Item = Page<E>;
-    type IntoIter = std::vec::IntoIter<Self::Item>;
-
-    fn into_iter(self) -> Self::IntoIter {
-        self.sheets.into_iter()
-    }
-}
-
-/// Implementation of [`Serialize`] for [`Book`] if the feature `serde` is enabled.
-#[cfg(feature = "serde")]
-impl<E> Serialize for Book<E>
-where
-    E: Serialize,
-{
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        #[derive(Serialize)]
-        struct BookModel<'a, E>
-        where
-            E: Serialize,
-        {
-            sheets: &'a Vec<Page<E>>,
-        }
-
-        let book_model: BookModel<E> = BookModel {
-            sheets: &self.sheets,
-        };
-
-        book_model.serialize(serializer)
-    }
-}
-
-/// Implementation of [`Deserialize`] for [`Book`] if the feature `serde` is enabled.
-#[cfg(feature = "serde")]
-impl<'de, E> DeDeserialize<'de> for Book<E>
-where
-    E: DeDeserialize<'de>,
-{
-    fn deserialize<D>(deserializer: D) -> Result<Book<E>, D::Error>
-    where
-        D: DeDeserializer<'de>,
-    {
-        #[derive(Deserialize)]
-        struct BookModel<E> {
-            sheets: Vec<Page<E>>,
-        }
-
-        let book_model: BookModel<E> = DeDeserialize::deserialize(deserializer)?;
-
-        Ok(Book {
-            sheets: book_model.sheets,
-        })
-    }
-}
-
-/// Implementation of [`ToSchema`] for [`Book`] if the feature `utoipa` is enabled.
-#[cfg(feature = "utoipa")]
-impl<'s, E> ToSchema<'s> for Book<E>
-where
-    E: ToSchema<'s>,
-{
-    fn schema() -> (&'s str, utoipa::openapi::RefOr<Schema>) {
-        (
-            "Book",
-            ObjectBuilder::new()
-                .description(Some("Model to represent a book of paginated items."))
-                .property(
-                    "sheets",
-                    ArrayBuilder::new()
-                        .description(Some(
-                            "Represents a paginated items as a collection of pages",
-                        ))
-                        .items(Page::<E>::schema().1),
-                )
-                .required("sheets")
-                .into(),
-        )
     }
 }
